[Update] typescript req passing error
-[Update] change the passing of the email from the JWT token from
req.email to res.locals.email

this error was issue due to a problem in the validator, passing data in
the request header was not working even if you create a interface to
make it work with it. saving the data in the request locals seemed to
solve this issue without any problems at all.

the middleware compilation problem is fixed now